requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
imutils==0.5.4
PyYAML==6.0.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
Flask==3.0.3
Werkzeug==3.0.3